Add more graceful exiting
I'm still instructing user to exit with ctrl+c so
I added keyboardInterrupt exception handling to make
exiting a bit nicer looking than the previous hassle.

Also discouraging user to use anything else than spacebar.
It proved a difficult task to clear the input buffer (?) that
is typed during playing the game. Currently using os.system('clear')
(I have a macbook) but that does not really work as intended. The
problem with the input buffer (?) is that whatever is typed during
the game is executed when game is exited, for example if user happens
to type 'exit' during the game, terminal is exited when game is stopped.

No good. Will have to think for a while. There might be something to
'ignore' all input but haven't looked into it yet.